requirements.txt
absl-py==0.7.0
appnope==0.1.0
args==0.1.0
asn1crypto==0.24.0
astor==0.7.1
awscli==1.16.52
backcall==0.1.0
backports-abc==0.5
backports.functools-lru-cache==1.5
backports.shutil-get-terminal-size==1.0.0
backports.weakref==1.0.post1
bcrypt==3.1.4
bleach==3.0.2
blessings==1.7
boto==2.49.0
botocore==1.12.42
Bottleneck==1.2.1
certifi==2018.10.15
cffi==1.11.5
chardet==3.0.4
Click==7.0
clint==0.5.1
cloudpickle==0.6.1
colorama==0.3.9
configparser==3.7.1
coverage==4.5.2
coveralls==1.5.1
cryptography==2.4.1
cycler==0.10.0
Cython==0.29.2
dask==0.20.2
decorator==4.3.0
defusedxml==0.5.0
docopt==0.6.2
docutils==0.14
entrypoints==0.2.3
enum34==1.1.6
fabric==2.4.0
funcsigs==1.0.2
functools32==3.2.3.post2
future==0.17.1
futures==3.2.0
gast==0.2.2
gpustat==0.5.0
graphviz==0.10.1
grpcio==1.19.0
h5py==2.8.0
hmmlearn==0.2.1
idna==2.6
invoke==1.2.0
ipaddress==1.0.22
ipdb==0.11
ipykernel==4.10.0
ipython==5.8.0
ipython-genutils==0.2.0
ipywidgets==7.4.2
jedi==0.13.1
Jinja2==2.10
jmespath==0.9.3
jsonschema==2.6.0
jupyter==1.0.0
jupyter-client==5.2.3
jupyter-console==5.2.0
jupyter-core==4.4.0
Keras==2.2.2
Keras-Applications==1.0.4
Keras-Preprocessing==1.0.2
kiwisolver==1.0.1
Lasagne==0.1
linecache2==1.0.0
Markdown==3.0.1
MarkupSafe==1.1.0
matplotlib==2.2.3
mistune==0.8.4
mock==2.0.0
nbconvert==5.4.0
nbformat==4.4.0
networkx==2.1
-e git+https://github.com/JackKelly/neuralnilm/@c0539416ddcdffc27e89fb393f31cdc0f437d50a#egg=NeuralNILM
-e git+https://github.com/nilmtk/nilm_metadata/@c7348dd1fdb2a7a7f9aa1f808ff5d67bf39f97c6#egg=nilm_metadata
-e git+https://github.com/nilmtk/nilmtk/@e5cc77e560d2520861a7ce3aec1f6feda4441c9e#egg=nilmtk
nltk==3.4
nose==1.3.7
notebook==5.7.0
numexpr==2.6.9
numpy==1.16.2
pandas==0.22.0
pandocfilters==1.4.2
paramiko==2.4.2
parso==0.3.1
pathlib2==2.3.3
pbr==5.1.3
pexpect==4.6.0
pickleshare==0.7.5
Pillow==5.3.0
pipenv==2018.10.13
plaidbench==0.3.5
plaidml==0.3.5
plaidml-keras==0.3.5
prettytable==0.7.2
prometheus-client==0.4.2
prompt-toolkit==1.0.15
protobuf==3.7.0
psutil==5.5.1
psycopg2==2.7.6.1
psycopg2-binary==2.7.6.1
ptyprocess==0.6.0
pyasn1==0.4.4
pycparser==2.19
pydot==1.3.0
Pygments==2.2.0
pymongo==3.7.2
PyNaCl==1.3.0
pyOpenSSL==19.0.0
pyparsing==2.3.0
python-dateutil==2.7.5
pytz==2018.7
PyWavelets==1.0.1
PyYAML==3.13
pyzmq==17.1.2
qtconsole==4.4.2
requests==2.18.4
rsa==3.4.2
s3transfer==0.1.13
scandir==1.9.0
scikit-image==0.14.1
scikit-learn==0.19.2
scipy==1.1.0
seaborn==0.9.0
Send2Trash==1.5.0
simplegeneric==0.8.1
singledispatch==3.4.0.3
six==1.12.0
sklearn==0.0
subprocess32==3.5.3
tables==3.4.4
tensorboard==1.12.0
tensorflow==1.12.0
tensorflow-estimator==1.13.0
termcolor==1.1.0
terminado==0.8.1
testpath==0.4.2
Theano==1.0.4
toolz==0.9.0
torch==1.0.0
torchvision==0.2.1
tornado==5.1.1
traceback2==1.4.0
traitlets==4.3.2
typing==3.6.6
unittest2==1.1.0
urllib3==1.22
virtualenv==16.3.0
virtualenv-clone==0.4.0
wcwidth==0.1.7
webencodings==0.5.1
Werkzeug==0.14.1
widgetsnbextension==3.4.2